refactor: Rename package service to internal
