+++ src/main/java/org/opentripplanner/service/worldenvelope/configure/WorldEnvelopeModule.java
@@ -2,8 +2,8 @@ package org.opentripplanner.service.worldenvelope.configure;
 
 import dagger.Binds;
 import dagger.Module;
+import org.opentripplanner.service.worldenvelope.internal.WorldEnvelopeRepository;
 import org.opentripplanner.service.worldenvelope.model.WorldEnvelopeService;
-import org.opentripplanner.service.worldenvelope.service.WorldEnvelopeModel;
 import org.opentripplanner.standalone.api.HttpRequestScoped;
 
 @Module
@@ -11,5 +11,5 @@ public abstract class WorldEnvelopeModule {
 
   @Binds
   @HttpRequestScoped
-  abstract WorldEnvelopeService bind(WorldEnvelopeModel service);
+  abstract WorldEnvelopeService bind(WorldEnvelopeRepository service);
 }
