+++ src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationFactory.java
@@ -8,7 +8,7 @@ import org.opentripplanner.raptor.configure.RaptorConfig;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripSchedule;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.service.worldenvelope.configure.WorldEnvelopeModule;
-import org.opentripplanner.service.worldenvelope.service.WorldEnvelopeModel;
+import org.opentripplanner.service.worldenvelope.internal.WorldEnvelopeRepository;
 import org.opentripplanner.standalone.api.OtpServerRequestContext;
 import org.opentripplanner.standalone.config.ConfigModel;
 import org.opentripplanner.standalone.config.configure.ConfigModule;
@@ -35,7 +35,7 @@ public interface ConstructApplicationFactory {
   RaptorConfig<TripSchedule> raptorConfig();
   Graph graph();
   TransitModel transitModel();
-  WorldEnvelopeModel worldEnvelopeModel();
+  WorldEnvelopeRepository worldEnvelopeModel();
 
   @Nullable
   GraphVisualizer graphVisualizer();
@@ -60,7 +60,7 @@ public interface ConstructApplicationFactory {
     Builder graphVisualizer(@Nullable GraphVisualizer graphVisualizer);
 
     @BindsInstance
-    Builder worldEnvelopeModel(WorldEnvelopeModel worldEnvelopeModel);
+    Builder worldEnvelopeModel(WorldEnvelopeRepository worldEnvelopeRepository);
 
     ConstructApplicationFactory build();
   }
