+++ src/main/java/org/opentripplanner/standalone/configure/LoadApplicationFactory.java
@@ -8,7 +8,7 @@ import org.opentripplanner.datastore.configure.DataStoreModule;
 import org.opentripplanner.ext.datastore.gs.GsDataSourceModule;
 import org.opentripplanner.graph_builder.GraphBuilderDataSources;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.service.worldenvelope.service.WorldEnvelopeModel;
+import org.opentripplanner.service.worldenvelope.internal.WorldEnvelopeRepository;
 import org.opentripplanner.standalone.config.CommandLineParameters;
 import org.opentripplanner.standalone.config.ConfigModel;
 import org.opentripplanner.standalone.config.configure.LoadConfigModule;
@@ -31,7 +31,7 @@ public interface LoadApplicationFactory {
   TransitModel emptyTransitModel();
 
   @Singleton
-  WorldEnvelopeModel emptyWorldWorldEnvelopeModel();
+  WorldEnvelopeRepository emptyWorldWorldEnvelopeModel();
 
   @Singleton
   GraphBuilderDataSources graphBuilderDataSources();
