+++ src/test/java/org/opentripplanner/TestServerContext.java
@@ -5,7 +5,7 @@ import static org.opentripplanner.standalone.configure.ConstructApplication.crea
 import io.micrometer.core.instrument.Metrics;
 import org.opentripplanner.raptor.configure.RaptorConfig;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.service.worldenvelope.service.WorldEnvelopeModel;
+import org.opentripplanner.service.worldenvelope.internal.WorldEnvelopeRepository;
 import org.opentripplanner.standalone.api.OtpServerRequestContext;
 import org.opentripplanner.standalone.config.RouterConfig;
 import org.opentripplanner.standalone.server.DefaultServerRequestContext;
@@ -32,7 +32,7 @@ public class TestServerContext {
       new DefaultTransitService(transitModel),
       Metrics.globalRegistry,
       routerConfig.vectorTileLayers(),
-      new WorldEnvelopeModel(),
+      new WorldEnvelopeRepository(),
       routerConfig.flexConfig(),
       null,
       routerConfig.requestLogFile()
