+++ src/test/java/org/opentripplanner/routing/graph/GraphSerializationTest.java
@@ -20,7 +20,7 @@ import org.opentripplanner.TestOtpModel;
 import org.opentripplanner.datastore.api.FileType;
 import org.opentripplanner.datastore.file.FileDataSource;
 import org.opentripplanner.framework.geometry.HashGridSpatialIndex;
-import org.opentripplanner.service.worldenvelope.service.WorldEnvelopeModel;
+import org.opentripplanner.service.worldenvelope.internal.WorldEnvelopeRepository;
 import org.opentripplanner.standalone.config.BuildConfig;
 import org.opentripplanner.standalone.config.RouterConfig;
 import org.opentripplanner.transit.model.framework.Deduplicator;
@@ -169,7 +169,7 @@ public class GraphSerializationTest {
     SerializedGraphObject serializedObj = new SerializedGraphObject(
       originalGraph,
       originalTransitModel,
-      new WorldEnvelopeModel(),
+      new WorldEnvelopeRepository(),
       BuildConfig.DEFAULT,
       RouterConfig.DEFAULT
     );
